web/requirements.txt
Django>=4.0
djangorestframework>=3.9
psycopg2-binary
python-dotenv
gunicorn
pillow
django-cors-headers
djangorestframework-simplejwt